# documentation-swagger/swagger.yaml
openapi: 3.0.1
info:
  version: '1.0.0'
  title: REST API Documentation
  description: REST API Documentation
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /users:
    get:
      tags:
        - user-controller
      description: Returns all users
      summary: Returns all users
      operationId: getAllUsers
      security:
        - BearerAdminAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
  /users/by-user-id-for-admin/{id}:
    get:
      tags:
        - user-controller
      description: Returns user by id
      summary: Returns all user by id
      security:
        - BearerAdminAuth: []
      operationId: getUserByIdForAdmin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
  /users/by-user-id:
    get:
      tags:
        - user-controller
      description: Returns user by id
      summary: Returns all user by id
      security:
        - BearerAdminAuth: []
      operationId: getUserById
      parameters:
        - name: User
          in: header
          description: Current authenticated user
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
  /orders:
    get:
      tags:
        - order-controller
      description: Returns all orders
      summary: Returns all orders
      security:
        - BearerAdminAuth: []
      operationId: getAllOrders
      parameters:
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orders'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
  /orders/by-user-id-for-admin/{user_id}:
    get:
      tags:
        - order-controller
      description: Returns order by user id
      summary: Returns order by user id
      security:
        - BearerAdminAuth: []
      operationId: getOrdersByUserIdForAdmin
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orders'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
  /orders/by-user-id:
    get:
      tags:
        - order-controller
      description: Returns orders by user id
      summary: Returns orders by user id
      security:
        - BearerUserOrAdminAuth: []
      operationId: getOrdersByUserId
      parameters:
        - name: User
          in: header
          description: Current authenticated user
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orders'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
  /orders/{giftCertificateId}:
    post:
      tags:
        - order-controller
      description: Returns created order
      summary: Returns created order
      security:
        - BearerUserOrAdminAuth: []
      operationId: createOrder
      parameters:
        - name: User
          in: header
          description: Current authenticated user
          required: true
          schema:
            type: integer
            format: int64
        - name: giftCertificateId
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
...
components:
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        _links:
          $ref: '#/components/schemas/Links'
    GiftCertificate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          minimum: 1
          type: integer
          format: int32
        duration:
          minimum: 1
          type: integer
          format: int32
        createDate:
          type: string
          format: date-time
        lastUpdateDate:
          type: string
          format: date-time
        tags:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        _links:
          $ref: '#/components/schemas/Links'
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        _links:
          $ref: '#/components/schemas/Links'
      required:
        - id
        - name
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cost:
          type: integer
          format: int128
        gift-certificate:
          type: array
          description: Order`s gift-certificate
          items:
            $ref: '#/components/schemas/GiftCertificate'
        user:
          type: array
          items:
            $ref: '#/components/schemas/User'
        purchaseDate:
          type: string
          format: date-time
        _links:
          $ref: '#/components/schemas/Links'
      required:
        - id
        - cost
        - gift-certificate
        - purchaseDate
...
    Users:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        _links:
          $ref: '#/components/schemas/Links'
    Orders:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        _links:
          $ref: '#/components/schemas/Links'
    HttpStatus400:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - BAD_REQUEST
        response body:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "BAD_REQUEST"
    HttpStatus401:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - Unauthorized
        response body:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    HttpStatus403:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - FORBIDDEN
        response body:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "FORBIDDEN"
    HttpStatus404:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - NOT_FOUND
        response body:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "NOT_FOUND"
...
    Links:
      type: array
      items:
        type: object
        properties:
          rel:
            type: string
          href:
            type: string
...
  securitySchemes:

    BearerAdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'For ADMIN'

    BearerUserOrAdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'For USER or ADMIN'

    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            email: See your primary Google Account email address
            profile: See your personal info, including any personal info you've made publicly available
            openid: Associate you with your personal info on Google